Code clean up, added/revised documentation, resolved issue #61.
