+++ tests/interfaces/encoded/data/JsonTestTrait.php
@@ -119,11 +119,69 @@ trait JsonTestTrait
         };
     }
 
+    /**
+     * Return true if the specified string is a valid json string,
+     * false otherwise.
+     *
+     * @return bool
+     *
+     * @example
+     *
+     * ```
+     * var_dump($this->stringIsAJsonString(
+     * '{"0":1,"1":[true,false],"foo":"bar"}'
+     * );
+     *
+     * // example output
+     * // true
+     *
+     * var_dump($this->stringIsAJsonString('abcdefg');
+     *
+     * // example output
+     * // false
+     *
+     * ```
+     *
+     */
     private function stringIsAJsonString(string $string): bool
     {
         return (false !== json_decode($string)) && (json_last_error() === JSON_ERROR_NONE);
     }
 
+    /**
+     * Return a json encoded version of specified string.
+     *
+     * If the string is already valid json string it will
+     * be returned unmodified.
+     *
+     * @return string
+     *
+     * @example
+     *
+     * ```
+     * var_dump($string);
+     *
+     * // example output:
+     * string(75) "Darling\PHPUnitTestUtilities\Tests\dev\mock\classes\ReflectedSubParentClass"
+     *
+     * var_dump($this->encodeStringAsJson($string));
+     *
+     * // example output:
+     * string(83) ""Darling\\PHPUnitTestUtilities\\Tests\\dev\\mock\\classes\\ReflectedSubParentClass""
+     *
+     * var_dump($string);
+     *
+     * // example output:
+     * string(38) "{"__class__":"stdClass","__data__":[]}"
+     *
+     * var_dump($this->encodeStringAsJson($string));
+     *
+     * // example output:
+     * string(38) "{"__class__":"stdClass","__data__":[]}"
+     *
+     * ```
+     *
+     */
     private function encodeStringAsJson(string $string): string
     {
         return match($this->stringIsAJsonString($string)) {
@@ -144,6 +202,55 @@ trait JsonTestTrait
      * @example
      *
      * ```
+     * var_dump($array);
+     *
+     * // example output:
+     * array(2) {
+     *   [0] =>
+     *   array(10) {
+     *     [0] =>
+     *     int(1)
+     *     [1] =>
+     *     bool(true)
+     *     [2] =>
+     *     bool(false)
+     *     [3] =>
+     *     NULL
+     *     [4] =>
+     *     string(6) "string"
+     *     [5] =>
+     *     array(0) {
+     *     }
+     *     [6] =>
+     *     class Darling\PHPTextTypes\classes\strings\Text#359 (1) {
+     *       private string $string =>
+     *       string(270) "zy@/d@u`^w:y7f5I$|*cg(ha/1Hl
+     * AW2=p`^8J9J@_2(</@zWd5g56q#4m+1qۗ4s<p=9Os5Awg#3op4!eWa$*7b%b0x6u^te!imoV=0>y59c^4%im^4&+&j=r^@,kf-;j8h3-)nZimZ4nfT_t"
+     *     }
+     *     'baz' =>
+     *     array(1) {
+     *       'secondary_id' =>
+     *       class Darling\PHPTextTypes\classes\strings\Id#360 (2) {
+     *         ...
+     *       }
+     *     }
+     *     'foo' =>
+     *     string(3) "bar"
+     *     'id' =>
+     *     class Darling\PHPTextTypes\classes\strings\Id#355 (2) {
+     *       private string $string =>
+     *       string(71) "QLmAr0bji5DOVPLxTUgsm0CV6zBzFFuaZfs9PtbbIZ57cJLV123gwBIGJPK8N6jMXeYi2yO"
+     *       private Darling\PHPTextTypes\interfaces\strings\Text $text =>
+     *       class Darling\PHPTextTypes\classes\strings\AlphanumericText#356 (2) {
+     *         ...
+     *       }
+     *     }
+     *   }
+     *
+     * var_dump($this->encodeArrayAsJson($array));
+     *
+     * // example output:
+     * string(1547) "{"0":1,"1":true,"2":false,"3":null,"4":"string","5":[],"6":"","baz":{"secondary_id":"{\"__class__\":\"Darling\\\\PHPTextTypes\\\\classes\\\\strings\\\\Id\",\"__data__\":{\"text\":\"{\\\"__class__\\\":\\\"Darling\\\\\\\\PHPTextTypes\\\\\\\\classes\\\\\\\\strings\\\\\\\\AlphanumericText\\\",\\\"__data__\\\":{\\\"text\\\":\\\"{\\\\\\\"__class__\\\\\\\":\\\\\\\"Darling\\\\\\\\\\\\\\\\PHPTextTypes\\\\\\\\\\\\\\\\classes\\\\\\\\\\\\\\\\strings\\\\\\\\\\\\\\\\Text\\\\\\\",\\\\\\\"__data__\\\\\\\":{\\\\\\\"string\\"...
      *
      * ```
      *
@@ -174,6 +281,33 @@ trait JsonTestTrait
      * @example
      *
      * ```
+     * var_dump($array);
+     *
+     * // example output:
+     * array(1) {
+     *   'secondary_id' =>
+     *   class Darling\PHPTextTypes\classes\strings\Id#360 (2) {
+     *     private string $string =>
+     *     string(77) "Mz5QlEwix4F56fMRzwDHOJtmBWdNhLN2Ax5XEQ7LBuq8NAKPlmsnwJ1K7l3gPn2ZUwQCIJYFKcckj"
+     *     private Darling\PHPTextTypes\interfaces\strings\Text $text =>
+     *     class Darling\PHPTextTypes\classes\strings\AlphanumericText#357 (2) {
+     *       private string $string =>
+     *       string(77) "Mz5QlEwix4F56fMRzwDHOJtmBWdNhLN2Ax5XEQ7LBuq8NAKPlmsnwJ1K7l3gPn2ZUwQCIJYFKcckj"
+     *       private Darling\PHPTextTypes\interfaces\strings\Text $text =>
+     *       class Darling\PHPTextTypes\classes\strings\Text#358 (1) {
+     *         ...
+     *       }
+     *     }
+     *   }
+     * }
+     *
+     * var_dump($this->encodeObjectsInArrayAsJson($array));
+     *
+     * // example output:
+     * array(1) {
+     *   'secondary_id' =>
+     *   string(595) "{"__class__":"Darling\\PHPTextTypes\\classes\\strings\\Id","__data__":{"text":"{\"__class__\":\"Darling\\\\PHPTextTypes\\\\classes\\\\strings\\\\AlphanumericText\",\"__data__\":{\"text\":\"{\\\"__class__\\\":\\\"Darling\\\\\\\\PHPTextTypes\\\\\\\\classes\\\\\\\\strings\\\\\\\\Text\\\",\\\"__data__\\\":{\\\"string\\\":\\\"Mz5QlEwix4F56fMRzwDHOJtmBWdNhLN2Ax5XEQ7LBuq8NAKPlmsnwJ1K7l3gPn2ZUwQCIJYFKcckj\\\"}}\",\"string\":\"Mz5QlEwix4F56fMRzwDHOJtmBWdNhLN2Ax5XEQ7LBuq8NAKPlmsnwJ1K7l3gPn2ZUwQCIJYFKcckj\"}}","string"...
+     * }
      *
      * ```
      *
@@ -193,6 +327,40 @@ trait JsonTestTrait
         return $array;
     }
 
+    /**
+     * Encode an object as json.
+     *
+     * @return string
+     *
+     * @example
+     *
+     * ```
+     * var_dump($object);
+     *
+     * // example output:
+     * class Darling\PHPTextTypes\classes\strings\Id#355 (2) {
+     *   private string $string =>
+     *   string(64) "1D7uY4fXgU5DM7I9dJ6gqN0FOM9CzImq8K00EAk3zmaLRw9ihCb3Jzsx1yY8oVnB"
+     *   private Darling\PHPTextTypes\interfaces\strings\Text $text =>
+     *   class Darling\PHPTextTypes\classes\strings\AlphanumericText#356 (2) {
+     *     private string $string =>
+     *     string(64) "1D7uY4fXgU5DM7I9dJ6gqN0FOM9CzImq8K00EAk3zmaLRw9ihCb3Jzsx1yY8oVnB"
+     *     private Darling\PHPTextTypes\interfaces\strings\Text $text =>
+     *     class Darling\PHPTextTypes\classes\strings\Text#354 (1) {
+     *       private string $string =>
+     *       string(64) "1D7uY4fXgU5DM7I9dJ6gqN0FOM9CzImq8K00EAk3zmaLRw9ihCb3Jzsx1yY8oVnB"
+     *     }
+     *   }
+     * }
+     *
+     * var_dump($this->encodeObjectAsJson($object));
+     *
+     * // example output:
+     * string(556) "{"__class__":"Darling\\PHPTextTypes\\classes\\strings\\Id","__data__":{"text":"{\"__class__\":\"Darling\\\\PHPTextTypes\\\\classes\\\\strings\\\\AlphanumericText\",\"__data__\":{\"text\":\"{\\\"__class__\\\":\\\"Darling\\\\\\\\PHPTextTypes\\\\\\\\classes\\\\\\\\strings\\\\\\\\Text\\\",\\\"__data__\\\":{\\\"string\\\":\\\"1D7uY4fXgU5DM7I9dJ6gqN0FOM9CzImq8K00EAk3zmaLRw9ihCb3Jzsx1yY8oVnB\\\"}}\",\"string\":\"1D7uY4fXgU5DM7I9dJ6gqN0FOM9CzImq8K00EAk3zmaLRw9ihCb3Jzsx1yY8oVnB\"}}","string":"1D7uY4fXgU5DM7I9dJ6gqN0"...
+     *
+     * ```
+     *
+     */
     private function encodeObjectAsJson(object $object): string
     {
         $data = [];
@@ -204,6 +372,9 @@ trait JsonTestTrait
                     $propertyValue
                 );
                continue;
+            }
+            if(is_array($propertyValue)) {
+                $propertyValue = $this->encodeObjectsInArrayAsJson($propertyValue);
             }
             $data[$propertyName] = $propertyValue;
 
@@ -220,6 +391,31 @@ trait JsonTestTrait
         );
     }
 
+    /**
+     * Return an instance of of an ObjectReflection that reflects
+     * the specified $object.
+     *
+     * @return ObjectReflection
+     *
+     * @example
+     *
+     * ```
+     * class Darling\PHPReflectionUtilities\classes\utilities\ObjectReflection#63 (2) {
+     *   private Darling\PHPTextTypes\classes\strings\ClassString $classString =>
+     *   class Darling\PHPTextTypes\classes\strings\ClassString#299 (1) {
+     *     private string $string =>
+     *     string(41) "Darling\PHPTextTypes\classes\strings\Text"
+     *   }
+     *   private object $object =>
+     *   class Darling\PHPTextTypes\classes\strings\Text#352 (1) {
+     *     private string $string =>
+     *     string(80) "RoLqEdBZfi8N0lTI6RQHwk5dF6MuMBnTWSJlCm3D6XnVwY8pc2YaEGPx1hjzWTJGSrg9WAd2IAw3iFHZ"
+     *   }
+     * }
+     *
+     * ```
+     *
+     */
     private function objectReflection(
         object $object
     ): ObjectReflection
